Move stream creation to Streaming plugin
