+++ lib/plugin/streaming.js
@@ -2,6 +2,7 @@ var Promise = require('bluebird');
 var util = require('util');
 var PluginAbstract = require('./abstract.js');
 var serviceLocator = require('../service-locator');
+var Stream = require('../stream');
 
 function PluginStreaming() {
   PluginStreaming.super_.apply(this, arguments);
@@ -17,6 +18,10 @@ util.inherits(PluginStreaming, PluginAbstract);
 PluginStreaming.prototype.processMessage = function(message) {
   var janusMessage = message['janus'];
 
+  if ('message' === janusMessage && 'create' === message['body']['request']) {
+    return this.onCreate(message);
+  }
+
   if ('webrtcup' === janusMessage) {
     return this.onWebrtcup(message);
   }
@@ -28,6 +33,31 @@ PluginStreaming.prototype.processMessage = function(message) {
   }
 
   return Promise.resolve();
+};
+
+/**
+ * @param {Object} message
+ * @returns {Promise}
+ */
+PluginStreaming.prototype.onCreate = function(message) {
+  var self = this;
+  this.proxyConnection.transactions.add(message['transaction'], function(response) {
+    if (self._isSuccessResponse(response)) {
+      self.stream = Stream.generate(message['body']['id'], self.proxyConnection);
+      return serviceLocator.get('cm-api-client').publish(self.stream.channelName, self.stream.id, Date.now(), self.proxyConnection.sessionData, message['data'])
+        .then(function() {
+          serviceLocator.get('logger').info('adding stream', self.stream);
+          serviceLocator.get('streams').add(self.stream);
+        })
+        .catch(function(error) {
+          self.stream.connection.close();
+          throw new JanusError.Error('Cannot publish: ', error);
+        });
+    } else {
+      return Promise.resolve();
+    }
+  });
+  return Promise.resolve(message);
 };
 
 /**
