+++ lib/plugin/video.js
@@ -20,39 +20,11 @@ PluginVideo.TYPE = 'janus.plugin.cm.rtpbroadcast';
 PluginVideo.prototype.processMessage = function(message) {
   var janusMessage = message['janus'];
 
-  if ('message' === janusMessage && 'create' === message['body']['request']) {
-    return this.onCreate(message);
-  }
   if ('message' === janusMessage && 'watch' === message['body']['request']) {
     return this.onWatch(message);
   }
 
   return PluginVideo.super_.prototype.processMessage.call(this, message);
-};
-
-/**
- * @param {Object} message
- * @returns {Promise}
- */
-PluginVideo.prototype.onCreate = function(message) {
-  var self = this;
-  this.proxyConnection.transactions.add(message['transaction'], function(response) {
-    if (self._isSuccessResponse(response)) {
-      self.stream = Stream.generate(message['body']['id'], self.proxyConnection);
-      return serviceLocator.get('cm-api-client').publish(self.stream.channelName, self.stream.id, Date.now(), self.proxyConnection.sessionData, message['data'])
-        .then(function() {
-          serviceLocator.get('logger').info('adding stream', self.stream);
-          serviceLocator.get('streams').add(self.stream);
-        })
-        .catch(function(error) {
-          self.stream.connection.close();
-          throw new JanusError.Error('Cannot publish: ', error);
-        });
-    } else {
-      return Promise.resolve();
-    }
-  });
-  return Promise.resolve(message);
 };
 
 /**
