+++ test/unit/plugin-streaming.js
@@ -74,6 +74,63 @@ describe('Streaming plugin', function() {
     assert(onHangupStub.calledWith(detachRequest));
   });
 
+  it('create stream', function(done) {
+    var proxyConnection = new ProxyConnection();
+    var plugin = new PluginStreaming('id', 'type', proxyConnection);
+    proxyConnection.plugins[plugin.id] = plugin;
+
+    var createRequest = {
+      janus: 'message',
+      body: {request: 'create', id: 'streamId'},
+      handle_id: plugin.id,
+      transaction: ProxyConnection.generateTransactionId()
+    };
+    var createResponse = {
+      janus: 'event',
+      plugindata: {data: {status: 'preparing'}},
+      handle_id: plugin.id,
+      transaction: createRequest.transaction
+    };
+
+    proxyConnection.processMessage(createRequest).then(function() {
+      proxyConnection.processMessage(createResponse).then(function() {
+        assert.equal(plugin.stream.channelName, createRequest.body.id);
+        var connectionStreams = serviceLocator.get('streams').findAllByConnection(proxyConnection);
+        assert.equal(connectionStreams.length, 1);
+        assert.equal(connectionStreams[0].channelName, createRequest.body.id);
+        done();
+      });
+    });
+  });
+
+  it('create stream fail', function(done) {
+    var proxyConnection = new ProxyConnection();
+    var plugin = new PluginStreaming('id', 'type', proxyConnection);
+    proxyConnection.plugins[plugin.id] = plugin;
+
+    var createRequest = {
+      janus: 'message',
+      body: {request: 'create', id: 'streamId'},
+      handle_id: plugin.id,
+      transaction: ProxyConnection.generateTransactionId()
+    };
+    var createResponse = {
+      janus: 'event',
+      plugindata: {data: {error: 'error'}},
+      handle_id: plugin.id,
+      transaction: createRequest.transaction
+    };
+
+    proxyConnection.processMessage(createRequest).then(function() {
+      proxyConnection.processMessage(createResponse).then(function() {
+        assert.isNull(plugin.stream);
+        var connectionStreams = serviceLocator.get('streams').findAllByConnection(proxyConnection);
+        assert.equal(connectionStreams.length, 0);
+        done();
+      });
+    });
+  });
+
   it('webrtcup', function(done) {
     var proxyConnection = new ProxyConnection();
     var plugin = new PluginStreaming('id', 'type', proxyConnection);
